+++ test/base_objectory_test.dart
@@ -3,10 +3,6 @@ import 'domain_model.dart';
 import 'package:test/test.dart';
 import 'package:bson/bson.dart';
 
-_addToObjectoryCache(PersistentObject obj) {
-  objectory.cache[obj.id.toString()] = obj;
-  obj.markAsFetched();
-}
 testAuthorCreation(){
   var author = new Author();
   author.name = 'vadim';
@@ -138,19 +134,19 @@ testObjectWithListtOfExternalRefs2Map() {
   father.firstName = 'Father';
   father.id = new ObjectId();
   father.map["_id"] = father.id;
-  _addToObjectoryCache(father);
+  objectory.addToCache(father);
   son = new Person();
   son.firstName = 'Son';
   son.father = father;
   son.id = new ObjectId();
   son.map["_id"] = son.id;
-  _addToObjectoryCache(son);
+  objectory.addToCache(son);
   daughter = new Person();
   daughter.father = father;
   daughter.firstName = 'daughter';
   daughter.id = new ObjectId();
   daughter.map["_id"] = daughter.id;
-  _addToObjectoryCache(daughter);
+  objectory.addToCache(daughter);
   father.children.add(son);
   father.children.add(null);
   father.children[1] = daughter;
@@ -163,7 +159,7 @@ testMap2ObjectWithListtOfInternalObjectsWithExternalRefs() {
   user.name = 'TestUser';
   user.id = new ObjectId();
   user.map["_id"] = user.id;
-  _addToObjectoryCache(user);
+  objectory.addToCache(user);
   Map articleMap = {"title": "test article", "body": "sasdfasdfasdf",
                     "comments": [{"body": "Excellent", "user": user.id}]};
   Article article = objectory.map2Object(Article,articleMap);